fix created product page
