+++ user-fe/src/pages/index.tsx
@@ -13,22 +13,10 @@ import { ProductFilterRecent } from "@/components/ProductFilterRecent";
 
 const inter = Inter({ subsets: ["latin"] });
 
-export default function Home({ productData }: { productData: ProductType[] }) {
-  console.log(productData);
+export default function Home() {
   return (
     <div>
-      {/* 
-      <Hero /> */}
-      {/* <RecommendedItems productData={productData} /> */}
-      <Product productData={productData} />
+      <a href="/productList">AWDAWDAWDAWDCD</a>
     </div>
   );
 }
-
-export const getServerSideProps = async () => {
-  const productRes = await instance.get("/getProducts");
-  const productData = productRes.data;
-  return {
-    props: { productData },
-  };
-};
